Refactor logger into modular subdirectory
