+++ src/ui/components/debug.rs
@@ -34,7 +34,7 @@ pub fn render_debug<B: Backend>(f: &mut Frame<B>, area: Rect, state: &AppState) 
 /// `logs/snapshots/` directory. The filename is timestamped using local time.
 pub fn write_debug_snapshot(state: &mut AppState) {
     let snapshot = DebugSnapshot::from_state(state);
-    match crate::logging::write_snapshot(&snapshot) {
+    match crate::logger::write_snapshot(&snapshot) {
         Ok(path) => {
             state.status_message = format!("Snapshot saved to {}", path.display());
             state.status_message_last_updated = Some(Instant::now());
